Add Bullets (#23)
* rename and change type Texture files

* update BulletState

* .

* .

* .

* .

* .

* .

* .

* .

+++ RifleManHideOnBushState.cpp
@@ -2,7 +2,7 @@
 
 RifleManHideOnBushState::RifleManHideOnBushState()
 {
-	this->time = GetTickCount64();
+	this->time = (FLOAT)GetTickCount64();
 }
 
 RifleManHideOnBushState::~RifleManHideOnBushState()
